+++ src/config.ts
@@ -1,0 +1,13 @@
+import type ConfigType from "./types/config";
+
+// Environment variables
+import dotenv from "dotenv";
+dotenv.config();
+
+const config:ConfigType = {
+	prefixEnabled: process.env.PREFIX_ENABLED == "true",
+	gptPrefix: "!gpt",
+	dallePrefix: "!dalle"
+}
+
+export { config };
