+++ src/helper/CONSTANT.ts
@@ -1,0 +1,3 @@
+export enum CONSTANT  {
+    StatusBroadcast = "status@broadcast"
+};
