+++ src/lib/whatsapp-client.ts
@@ -1,0 +1,9 @@
+import {  Client } from "whatsapp-web.js";
+
+// Whatsapp Client
+export const whatsappClient = new Client({
+	puppeteer: {
+		args: ["--no-sandbox"]
+	}
+});
+
