+++ src/models/dalle.ts
