+++ src/models/gpt.ts
